Support non greedy operator in route variables (#5)
This adds support in route variables to not be greedy in the way they consume input.  By default <foo> will turn into a (?P<foo>.+) regex which will consume as many characters as possible in matching. There are cases with multiple variables where you want it to not operate greedily. This introduces '?' (the python non greedy operator) to get the behaviour of turning <foo?> into (?P<foo>.+?) which will stop at first match that works rather than greedy match.  See added tests for an example of behavior.

Example:

'<a> to <b>' and parsing 'a to b to c'.
With greedy, you will get a = 'a to b', and b = 'c'.
With non greedy ('<a?> to <b>'), you will get a = 'a', and b = 'b to c'