[INTPROD-1125] Update Runner for GH Action (#20)
This pull request primarily focuses on updating the GitHub Actions
workflows and the pre-commit configuration. The updates include changing
the operating system used in the workflows from `ubuntu-18.04` to
`ubuntu-latest`, upgrading the Python version from `3.6` to `3.8`, and
updating the `flake8` repository and revision in the pre-commit
configuration.

GitHub Actions workflows:

*
[`.github/workflows/pull_request.yml`](diffhunk://#diff-a0fe23534b616d51ce686d2a1bcd1a78bc75074aef1a2f6ee96c9469991e1a4cL4-R20):
Updated the operating system from `ubuntu-18.04` to `ubuntu-latest` and
upgraded the Python version from `3.6` to `3.8`. Also, removed the
support for Python 2.x in the test matrix.
*
[`.github/workflows/push.yml`](diffhunk://#diff-f3fc934cf0d89bdf07de358896ff90f1202585df812cf606206d1830a9949811L11-R18):
In both the `build-and-deploy-docs` and
`build-and-publish-python-module` jobs, updated the operating system
from `ubuntu-18.04` to `ubuntu-latest` and upgraded the Python version
from `3.6` to `3.8`.
[[1]](diffhunk://#diff-f3fc934cf0d89bdf07de358896ff90f1202585df812cf606206d1830a9949811L11-R18)
[[2]](diffhunk://#diff-f3fc934cf0d89bdf07de358896ff90f1202585df812cf606206d1830a9949811L34-R41)

Pre-commit configuration:

*
[`.pre-commit-config.yaml`](diffhunk://#diff-63a9c44a44acf85fea213a857769990937107cf072831e1a26808cfde9d096b9L3-R4):
Updated the `flake8` repository from `https://gitlab.com/pycqa/flake8`
to `https://github.com/pycqa/flake8` and updated the revision from
`3.7.9` to `5.0.4`.